Split App.js into some seperate files. This is a start to make the code more manageable.
